+++ test/keypin/core_test.clj
@@ -218,7 +218,9 @@
   (is (= [{:a "foo" :b "10"}
           {:a "bar" :b "20" :c "22"}
           {:a "baz" :b "30" :c "32" :d "34"}]
-        (ku/str->tuples [:a :b :c :d] "foo" "foo: 10, bar: 20: 22, baz: 30: 32: 34: 36"))))
+        (ku/str->tuples [:a :b :c :d] "foo" "foo: 10, bar: 20: 22, baz: 30: 32: 34: 36")))
+  (is (= [:foo 100]
+        (ku/str->edn "foo" "[:foo 100]"))))
 
 
 (deftest test-optional-parsers
@@ -276,7 +278,11 @@
   (is (= [{:a "foo" :b "10"}
           {:a "bar" :b "20" :c "22"}
           {:a "baz" :b "30" :c "32" :d "34"}]
-        (ku/any->tuples [:a :b :c :d] "foo" "foo: 10, bar: 20: 22, baz: 30: 32: 34: 36"))))
+        (ku/any->tuples [:a :b :c :d] "foo" "foo: 10, bar: 20: 22, baz: 30: 32: 34: 36")))
+  (is (= [:foo 100]
+        (ku/any->edn "foo" "[:foo 100]")))
+  (is (= [:foo 100]
+        (ku/any->edn "foo" [:foo 100]))))
 
 
 (defkey